requirements.txt
ultralytics
boto3
Pillow==9.5.0